+++ src/main/java/com/example/studentselectionsystem/repository/UserRepository.java
@@ -52,7 +52,7 @@ public interface UserRepository extends BaseMapper<User> {
      * @return 用户信息
      */
     @Select("SELECT u.*, r.* FROM users u LEFT JOIN user_role ur ON u.id = ur.user_id LEFT JOIN role r ON ur.role_id = r.id WHERE u.id = #{id}")
-    Optional<User> selectByIdWithRoles(@Param("id") Integer id);
+    Optional<User> selectByIdWithRoles(@Param("id") Long id);
     
     /**
      * 根据真实姓名查找用户
